[addon] Move add-on library parts to independent folders (code rework)
