+++ Makefile.in
@@ -25,6 +25,14 @@ VideoPlayer_ARCHIVES=xbmc/cores/VideoPlayer/VideoPlayer.a \
 
 DIRECTORY_ARCHIVES=$(VideoPlayer_ARCHIVES) \
                    xbmc/addons/addons.a \
+                   xbmc/addons/binary/interfaces/addon-interfaces.a \
+                   xbmc/addons/binary/interfaces/api1/Addon/addon-callbacks-addon.a \
+                   xbmc/addons/binary/interfaces/api1/AudioDSP/addon-callbacks-audiodsp.a \
+                   xbmc/addons/binary/interfaces/api1/AudioEngine/addon-callbacks-audioengine.a \
+                   xbmc/addons/binary/interfaces/api1/Codec/addon-callbacks-codec.a \
+                   xbmc/addons/binary/interfaces/api1/GUI/addon-callbacks-gui.a \
+                   xbmc/addons/binary/interfaces/api1/InputStream/addon-callbacks-inputstream.a \
+                   xbmc/addons/binary/interfaces/api1/PVR/addon-callbacks-pvr.a \
                    xbmc/cdrip/cdrip.a \
                    xbmc/contrib/kissfft/kissfft.a \
                    xbmc/cores/AudioEngine/audioengine.a \
