+++ lib/addons/library.kodi.adsp/libKODI_adsp.cpp
@@ -23,7 +23,7 @@
 #include <stdarg.h>
 #include <string>
 #include "addons/kodi-addon-dev-kit/include/kodi/libKODI_adsp.h"
-#include "addons/AddonCallbacks.h"
+#include "addons/binary/interfaces/api1/AudioDSP/AddonCallbacksAudioDSP.h"
 
 #ifdef _WIN32
 #include <windows.h>
@@ -33,6 +33,7 @@
 #endif
 
 using namespace std;
+using namespace V1::KodiAPI::AudioDSP;
 
 extern "C"
 {
@@ -44,7 +45,7 @@ DLLEXPORT void* ADSP_register_me(void *hdl)
     fprintf(stderr, "libKODI_adsp-ERROR: ADSPLib_register_me is called with NULL handle !!!\n");
   else
   {
-    cb = ((AddonCB*)hdl)->ADSPLib_RegisterMe(((AddonCB*)hdl)->addonData);
+    cb = (CB_ADSPLib*)((AddonCB*)hdl)->ADSPLib_RegisterMe(((AddonCB*)hdl)->addonData);
     if (!cb)
       fprintf(stderr, "libKODI_adsp-ERROR: ADSPLib_register_me can't get callback table from KODI !!!\n");
   }
