+++ lib/addons/library.xbmc.codec/libXBMC_codec.cpp
@@ -22,7 +22,7 @@
 #include <stdlib.h>
 #include <stdarg.h>
 #include <string>
-#include "addons/AddonCallbacks.h"
+#include "addons/binary/interfaces/api1/Codec/AddonCallbacksCodec.h"
 #include "addons/kodi-addon-dev-kit/include/kodi/libXBMC_codec.h"
 
 #ifdef _WIN32
@@ -33,18 +33,19 @@
 #endif
 
 using namespace std;
+using namespace V1::KodiAPI::Codec;
 
 extern "C"
 {
 
 DLLEXPORT void* CODEC_register_me(void *hdl)
 {
-  CB_CODECLib *cb = NULL;
+  CB_CodecLib *cb = NULL;
   if (!hdl)
     fprintf(stderr, "libXBMC_codec-ERROR: %s is called with NULL handle\n", __FUNCTION__);
   else
   {
-    cb = ((AddonCB*)hdl)->CODECLib_RegisterMe(((AddonCB*)hdl)->addonData);
+    cb = (CB_CodecLib*)((AddonCB*)hdl)->CodecLib_RegisterMe(((AddonCB*)hdl)->addonData);
     if (!cb)
       fprintf(stderr, "libXBMC_codec-ERROR: %s can't get callback table from XBMC\n", __FUNCTION__);
   }
@@ -54,7 +55,7 @@ DLLEXPORT void* CODEC_register_me(void *hdl)
 DLLEXPORT void CODEC_unregister_me(void *hdl, void* cb)
 {
   if (hdl && cb)
-    ((AddonCB*)hdl)->CODECLib_UnRegisterMe(((AddonCB*)hdl)->addonData, (CB_CODECLib*)cb);
+    ((AddonCB*)hdl)->CodecLib_UnRegisterMe(((AddonCB*)hdl)->addonData, (CB_CodecLib*)cb);
 }
 
 DLLEXPORT xbmc_codec_t CODEC_get_codec_by_name(void *hdl, void* cb, const char* strCodecName)
@@ -64,7 +65,7 @@ DLLEXPORT xbmc_codec_t CODEC_get_codec_by_name(void *hdl, void* cb, const char* 
   retVal.codec_type = XBMC_CODEC_TYPE_UNKNOWN;
 
   if (cb != NULL)
-    retVal = ((CB_CODECLib*)cb)->GetCodecByName(((AddonCB*)hdl)->addonData, strCodecName);
+    retVal = ((CB_CodecLib*)cb)->GetCodecByName(((AddonCB*)hdl)->addonData, strCodecName);
 
   return retVal;
 }
