+++ project/cmake/treedata/common/subdirs.txt
@@ -1,25 +1,33 @@
-xbmc                            xbmc
-xbmc/addons                     addons
-xbmc/commons                    commons
-xbmc/cdrip                      cdrip
-xbmc/dbwrappers                 dbwrappers
-xbmc/dialogs                    dialogs
-xbmc/epg                        epg
-xbmc/guilib                     guilib
-xbmc/input                      input
-xbmc/listproviders              listproviders
-xbmc/media                      media
-xbmc/messaging                  messaging
-xbmc/messaging/helpers          messagingHelpers
-xbmc/pictures                   pictures
-xbmc/platform                   platform
-xbmc/playlists                  playlists
-xbmc/powermanagement            powermanagement
-xbmc/programs                   programs
-xbmc/rendering                  rendering
-xbmc/storage                    storage
-xbmc/threads                    threads
-xbmc/utils                      utils
-xbmc/view                       view
-xbmc/windowing                  windowing
-xbmc/windows                    windows
+xbmc                                            xbmc
+xbmc/addons                                     addons
+xbmc/addons/binary/interfaces                   addonsBinaryInterfaces
+xbmc/addons/binary/interfaces/api1/Addon        api1AddonCallbacks_Addon
+xbmc/addons/binary/interfaces/api1/AudioDSP     api1AddonCallbacks_AudioDSP
+xbmc/addons/binary/interfaces/api1/AudioEngine  api1AddonCallbacks_AudioEngine
+xbmc/addons/binary/interfaces/api1/Codec        api1AddonCallbacks_Codec
+xbmc/addons/binary/interfaces/api1/GUI          api1AddonCallbacks_GUI
+xbmc/addons/binary/interfaces/api1/InputStream  api1AddonCallbacks_InputStream
+xbmc/addons/binary/interfaces/api1/PVR          api1AddonCallbacks_PVR
+xbmc/commons                                    commons
+xbmc/cdrip                                      cdrip
+xbmc/dbwrappers                                 dbwrappers
+xbmc/dialogs                                    dialogs
+xbmc/epg                                        epg
+xbmc/guilib                                     guilib
+xbmc/input                                      input
+xbmc/listproviders                              listproviders
+xbmc/media                                      media
+xbmc/messaging                                  messaging
+xbmc/messaging/helpers                          messagingHelpers
+xbmc/pictures                                   pictures
+xbmc/platform                                   platform
+xbmc/playlists                                  playlists
+xbmc/powermanagement                            powermanagement
+xbmc/programs                                   programs
+xbmc/rendering                                  rendering
+xbmc/storage                                    storage
+xbmc/threads                                    threads
+xbmc/utils                                      utils
+xbmc/view                                       view
+xbmc/windowing                                  windowing
+xbmc/windows                                    windows
