+++ xbmc/addons/Makefile
@@ -1,13 +1,5 @@
 SRCS=Addon.cpp \
      AddonBuilder.cpp \
-     AddonCallbacks.cpp \
-     AddonCallbacksAddon.cpp \
-     AddonCallbacksAudioDSP.cpp \
-     AddonCallbacksAudioEngine.cpp \
-     AddonCallbacksCodec.cpp \
-     AddonCallbacksGUI.cpp \
-     AddonCallbacksPVR.cpp \
-     AddonCallbacksInputStream.cpp \
      AddonDatabase.cpp \
      AddonInstaller.cpp \
      AddonManager.cpp \
