+++ xbmc/addons/binary/interfaces/CMakeLists.txt
@@ -1,0 +1,10 @@
+set(SOURCES AddonInterfaces.cpp)
+
+set(HEADERS AddonInterfaces.h
+            IAddonInterface.h)
+
+core_add_library(addonsBinaryInterfaces)
+
+if(ENABLE_INTERNAL_FFMPEG)
+  add_dependencies(addons ffmpeg)
+endif()
