+++ xbmc/addons/binary/interfaces/api1/Addon/AddonCallbacksAddon.h
@@ -1,7 +1,8 @@
 #pragma once
 /*
  *      Copyright (C) 2012-2013 Team XBMC
- *      http://xbmc.org
+ *      Copyright (C) 2015-2016 Team KODI
+ *      http://kodi.tv
  *
  *  This Program is free software; you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -14,32 +15,126 @@
  *  GNU General Public License for more details.
  *
  *  You should have received a copy of the GNU General Public License
- *  along with XBMC; see the file COPYING.  If not, see
+ *  along with KODI; see the file COPYING.  If not, see
  *  <http://www.gnu.org/licenses/>.
  *
  */
 
-#include "AddonCallbacks.h"
+#include "addons/binary/interfaces/AddonInterfaces.h"
+#include "addons/kodi-addon-dev-kit/include/kodi/libXBMC_addon.h"
 
 struct VFSDirEntry;
 
+#ifdef TARGET_WINDOWS
+#ifndef _SSIZE_T_DEFINED
+typedef intptr_t ssize_t;
+#define _SSIZE_T_DEFINED
+#endif // !_SSIZE_T_DEFINED
+#endif // TARGET_WINDOWS
+
 namespace ADDON
 {
+  class CAddon;
+};
 
-class CAddonCallbacksAddon
+namespace V1
+{
+namespace KodiAPI
+{
+
+namespace AddOn
+{
+
+typedef void (*AddOnLogCallback)(void *addonData, const ADDON::addon_log_t loglevel, const char *msg);
+typedef void (*AddOnQueueNotification)(void *addonData, const ADDON::queue_msg_t type, const char *msg);
+typedef bool (*AddOnWakeOnLan)(const char* mac);
+typedef bool (*AddOnGetSetting)(void *addonData, const char *settingName, void *settingValue);
+typedef char* (*AddOnUnknownToUTF8)(const char *sourceDest);
+typedef char* (*AddOnGetLocalizedString)(const void* addonData, long dwCode);
+typedef char* (*AddOnGetDVDMenuLanguage)(const void* addonData);
+typedef void (*AddOnFreeString)(const void* addonData, char* str);
+
+typedef void* (*AddOnOpenFile)(const void* addonData, const char* strFileName, unsigned int flags);
+typedef void* (*AddOnOpenFileForWrite)(const void* addonData, const char* strFileName, bool bOverWrite);
+typedef ssize_t (*AddOnReadFile)(const void* addonData, void* file, void* lpBuf, size_t uiBufSize);
+typedef bool (*AddOnReadFileString)(const void* addonData, void* file, char *szLine, int iLineLength);
+typedef ssize_t (*AddOnWriteFile)(const void* addonData, void* file, const void* lpBuf, size_t uiBufSize);
+typedef void (*AddOnFlushFile)(const void* addonData, void* file);
+typedef int64_t (*AddOnSeekFile)(const void* addonData, void* file, int64_t iFilePosition, int iWhence);
+typedef int (*AddOnTruncateFile)(const void* addonData, void* file, int64_t iSize);
+typedef int64_t (*AddOnGetFilePosition)(const void* addonData, void* file);
+typedef int64_t (*AddOnGetFileLength)(const void* addonData, void* file);
+typedef double (*AddOnGetFileDownloadSpeed)(const void* addonData, void* file);
+typedef void (*AddOnCloseFile)(const void* addonData, void* file);
+typedef int (*AddOnGetFileChunkSize)(const void* addonData, void* file);
+typedef bool (*AddOnFileExists)(const void* addonData, const char *strFileName, bool bUseCache);
+typedef int (*AddOnStatFile)(const void* addonData, const char *strFileName, struct __stat64* buffer);
+typedef bool (*AddOnDeleteFile)(const void* addonData, const char *strFileName);
+typedef bool (*AddOnCanOpenDirectory)(const void* addonData, const char* strURL);
+typedef bool (*AddOnCreateDirectory)(const void* addonData, const char *strPath);
+typedef bool (*AddOnDirectoryExists)(const void* addonData, const char *strPath);
+typedef bool (*AddOnRemoveDirectory)(const void* addonData, const char *strPath);
+typedef bool (*AddOnGetDirectory)(const void* addonData, const char *strPath, const char* mask, VFSDirEntry** items, unsigned int* num_items);
+typedef void (*AddOnFreeDirectory)(const void* addonData, VFSDirEntry* items, unsigned int num_items);
+typedef void* (*AddOnCURLCreate)(const void* addonData, const char* strURL);
+typedef bool (*AddOnCURLAddOption)(const void* addonData, void* file, XFILE::CURLOPTIONTYPE type, const char* name, const char * value);
+typedef bool (*AddOnCURLOpen)(const void* addonData, void* file, unsigned int flags);
+
+typedef struct CB_AddOn
+{
+  AddOnLogCallback            Log;
+  AddOnQueueNotification      QueueNotification;
+  AddOnWakeOnLan              WakeOnLan;
+  AddOnGetSetting             GetSetting;
+  AddOnUnknownToUTF8          UnknownToUTF8;
+  AddOnGetLocalizedString     GetLocalizedString;
+  AddOnGetDVDMenuLanguage     GetDVDMenuLanguage;
+  AddOnFreeString             FreeString;
+
+  AddOnOpenFile               OpenFile;
+  AddOnOpenFileForWrite       OpenFileForWrite;
+  AddOnReadFile               ReadFile;
+  AddOnReadFileString         ReadFileString;
+  AddOnWriteFile              WriteFile;
+  AddOnFlushFile              FlushFile;
+  AddOnSeekFile               SeekFile;
+  AddOnTruncateFile           TruncateFile;
+  AddOnGetFilePosition        GetFilePosition;
+  AddOnGetFileLength          GetFileLength;
+  AddOnGetFileDownloadSpeed   GetFileDownloadSpeed;
+  AddOnCloseFile              CloseFile;
+  AddOnGetFileChunkSize       GetFileChunkSize;
+  AddOnFileExists             FileExists;
+  AddOnStatFile               StatFile;
+  AddOnDeleteFile             DeleteFile;
+  AddOnCanOpenDirectory       CanOpenDirectory;
+  AddOnCreateDirectory        CreateDirectory;
+  AddOnDirectoryExists        DirectoryExists;
+  AddOnRemoveDirectory        RemoveDirectory;
+  AddOnGetDirectory           GetDirectory;
+  AddOnFreeDirectory          FreeDirectory;
+  AddOnCURLCreate             CURLCreate;
+  AddOnCURLAddOption          CURLAddOption;
+  AddOnCURLOpen               CURLOpen;
+} CB_AddOnLib;
+
+class CAddonCallbacksAddon : public ADDON::IAddonInterface
 {
 public:
-  CAddonCallbacksAddon(CAddon* addon);
-  ~CAddonCallbacksAddon();
+  CAddonCallbacksAddon(ADDON::CAddon* addon);
+  virtual ~CAddonCallbacksAddon();
+
+  static int APILevel()        { return 1; }
+  static std::string Version() { return "0.0.1"; }
 
   /*!
    * @return The callback table.
    */
   CB_AddOnLib *GetCallbacks() { return m_callbacks; }
 
-  static void AddOnLog(void *addonData, const addon_log_t addonLogLevel, const char *strMessage);
+  static void AddOnLog(void *addonData, const ADDON::addon_log_t addonLogLevel, const char *strMessage);
   static bool GetAddonSetting(void *addonData, const char *strSettingName, void *settingValue);
-  static void QueueNotification(void *addonData, const queue_msg_t type, const char *strMessage);
+  static void QueueNotification(void *addonData, const ADDON::queue_msg_t type, const char *strMessage);
   static bool WakeOnLan(const char *mac);
   static char* UnknownToUTF8(const char *strSource);
   static char* GetLocalizedString(const void* addonData, long dwCode);
@@ -75,7 +170,9 @@ public:
 
 private:
   CB_AddOnLib  *m_callbacks; /*!< callback addresses */
-  CAddon       *m_addon;     /*!< the add-on */
 };
 
-}; /* namespace ADDON */
+}; /* namespace AddOn */
+
+}; /* namespace KodiAPI */
+}; /* namespace V1 */
