+++ xbmc/addons/binary/interfaces/api1/Addon/Makefile
@@ -1,0 +1,7 @@
+SRCS=AddonCallbacksAddon.cpp \
+
+LIB=addon-callbacks-addon.a
+
+include ../../../../../../Makefile.include
+-include $(patsubst %.cpp,%.P,$(patsubst %.c,%.P,$(SRCS)))
+
