+++ xbmc/addons/binary/interfaces/api1/AudioDSP/Makefile
@@ -1,0 +1,7 @@
+SRCS=AddonCallbacksAudioDSP.cpp \
+
+LIB=addon-callbacks-audiodsp.a
+
+include ../../../../../../Makefile.include
+-include $(patsubst %.cpp,%.P,$(patsubst %.c,%.P,$(SRCS)))
+
