+++ xbmc/addons/binary/interfaces/api1/AudioEngine/AddonCallbacksAudioEngine.h
@@ -1,6 +1,6 @@
 #pragma once
 /*
- *      Copyright (C) 2014 Team KODI
+ *      Copyright (C) 2014-2016 Team KODI
  *      http://kodi.tv
  *
  *  This Program is free software; you can redistribute it and/or modify
@@ -19,20 +19,91 @@
  *
  */
 
-#include "AddonCallbacks.h"
+#include "addons/binary/interfaces/AddonInterfaces.h"
+#include "addons/kodi-addon-dev-kit/include/kodi/libKODI_audioengine.h"
 #include "addons/kodi-addon-dev-kit/include/kodi/kodi_audioengine_types.h"
 
-namespace ADDON
+namespace V1
 {
+namespace KodiAPI
+{
+
+namespace AudioEngine
+{
+
+// ---------------------------------------
+// libKODI_audioengine definitions
+// ---------------------------------------
+typedef AEStreamHandle* (*AudioEngine_MakeStream)(AudioEngineFormat Format, unsigned int Options);
+typedef void            (*AudioEngine_FreeStream)(AEStreamHandle *stream);
+typedef bool            (*AudioEngine_GetCurrentSinkFormat)(void *addonData, AudioEngineFormat *SinkFormat);
+
+// Audio Engine Stream definitions
+typedef unsigned int            (*AudioEngine_Stream_GetSpace)(void *addonData, AEStreamHandle *handle);
+typedef unsigned int            (*AudioEngine_Stream_AddData)(void *addonData, AEStreamHandle *handle, uint8_t* const *Data, unsigned int Offset, unsigned int Frames);
+typedef double                  (*AudioEngine_Stream_GetDelay)(void *addonData, AEStreamHandle *handle);
+typedef bool                    (*AudioEngine_Stream_IsBuffering)(void *addonData, AEStreamHandle *handle);
+typedef double                  (*AudioEngine_Stream_GetCacheTime)(void *addonData, AEStreamHandle *handle);
+typedef double                  (*AudioEngine_Stream_GetCacheTotal)(void *addonData, AEStreamHandle *handle);
+typedef void                    (*AudioEngine_Stream_Pause)(void *addonData, AEStreamHandle *handle);
+typedef void                    (*AudioEngine_Stream_Resume)(void *addonData, AEStreamHandle *handle);
+typedef void                    (*AudioEngine_Stream_Drain)(void *addonData, AEStreamHandle *handle, bool Wait);
+typedef bool                    (*AudioEngine_Stream_IsDraining)(void *addonData, AEStreamHandle *handle);
+typedef bool                    (*AudioEngine_Stream_IsDrained)(void *addonData, AEStreamHandle *handle);
+typedef void                    (*AudioEngine_Stream_Flush)(void *addonData, AEStreamHandle *handle);
+typedef float                   (*AudioEngine_Stream_GetVolume)(void *addonData, AEStreamHandle *handle);
+typedef void                    (*AudioEngine_Stream_SetVolume)(void *addonData, AEStreamHandle *handle, float Volume);
+typedef float                   (*AudioEngine_Stream_GetAmplification)(void *addonData, AEStreamHandle *handle);
+typedef void                    (*AudioEngine_Stream_SetAmplification)(void *addonData, AEStreamHandle *handle, float Amplify);
+typedef const unsigned int      (*AudioEngine_Stream_GetFrameSize)(void *addonData, AEStreamHandle *handle);
+typedef const unsigned int      (*AudioEngine_Stream_GetChannelCount)(void *addonData, AEStreamHandle *handle);
+typedef const unsigned int      (*AudioEngine_Stream_GetSampleRate)(void *addonData, AEStreamHandle *handle);
+typedef const AEDataFormat      (*AudioEngine_Stream_GetDataFormat)(void *addonData, AEStreamHandle *handle);
+typedef double                  (*AudioEngine_Stream_GetResampleRatio)(void *addonData, AEStreamHandle *handle);
+typedef void                    (*AudioEngine_Stream_SetResampleRatio)(void *addonData, AEStreamHandle *handle, double Ratio);
+
+typedef struct CB_AudioEngineLib
+{
+  AudioEngine_MakeStream                    MakeStream;
+  AudioEngine_FreeStream                    FreeStream;
+  AudioEngine_GetCurrentSinkFormat          GetCurrentSinkFormat;
+
+  // AudioEngine stream callbacks
+  AudioEngine_Stream_GetSpace               AEStream_GetSpace;
+  AudioEngine_Stream_AddData                AEStream_AddData;
+  AudioEngine_Stream_GetDelay               AEStream_GetDelay;
+  AudioEngine_Stream_IsBuffering            AEStream_IsBuffering;
+  AudioEngine_Stream_GetCacheTime           AEStream_GetCacheTime;
+  AudioEngine_Stream_GetCacheTotal          AEStream_GetCacheTotal;
+  AudioEngine_Stream_Pause                  AEStream_Pause;
+  AudioEngine_Stream_Resume                 AEStream_Resume;
+  AudioEngine_Stream_Drain                  AEStream_Drain;
+  AudioEngine_Stream_IsDraining             AEStream_IsDraining;
+  AudioEngine_Stream_IsDrained              AEStream_IsDrained;
+  AudioEngine_Stream_Flush                  AEStream_Flush;
+  AudioEngine_Stream_GetVolume              AEStream_GetVolume;
+  AudioEngine_Stream_SetVolume              AEStream_SetVolume;
+  AudioEngine_Stream_GetAmplification       AEStream_GetAmplification;
+  AudioEngine_Stream_SetAmplification       AEStream_SetAmplification;
+  AudioEngine_Stream_GetFrameSize           AEStream_GetFrameSize;
+  AudioEngine_Stream_GetChannelCount        AEStream_GetChannelCount;
+  AudioEngine_Stream_GetSampleRate          AEStream_GetSampleRate;
+  AudioEngine_Stream_GetDataFormat          AEStream_GetDataFormat;
+  AudioEngine_Stream_GetResampleRatio       AEStream_GetResampleRatio;
+  AudioEngine_Stream_SetResampleRatio       AEStream_SetResampleRatio;
+} CB_AudioEngineLib;
 
 /*!
  * Callbacks for Kodi's AudioEngine.
  */
-class CAddonCallbacksAudioEngine
+class CAddonCallbacksAudioEngine : public ADDON::IAddonInterface
 {
 public:
-  CAddonCallbacksAudioEngine(CAddon* Addon);
-  ~CAddonCallbacksAudioEngine();
+  CAddonCallbacksAudioEngine(ADDON::CAddon* Addon);
+  virtual ~CAddonCallbacksAudioEngine();
+
+  static int APILevel() { return 1; }
+  static std::string Version() { return KODI_AUDIOENGINE_API_VERSION; }
 
   /*!
    * @return The callback table.
@@ -200,7 +271,9 @@ public:
 
 private:
   CB_AudioEngineLib   *m_callbacks; /*!< callback addresses */
-  CAddon              *m_addon;     /*!< the addon */
 };
 
-}; /* namespace ADDON */
+}; /* namespace AudioEngine */
+
+}; /* namespace KodiAPI */
+}; /* namespace V1 */
