+++ xbmc/addons/binary/interfaces/api1/AudioEngine/Makefile
@@ -1,0 +1,6 @@
+SRCS=AddonCallbacksAudioEngine.cpp \
+
+LIB=addon-callbacks-audioengine.a
+
+include ../../../../../../Makefile.include
+-include $(patsubst %.cpp,%.P,$(patsubst %.c,%.P,$(SRCS)))
