+++ xbmc/addons/binary/interfaces/api1/Codec/Makefile
@@ -1,0 +1,7 @@
+SRCS=AddonCallbacksCodec.cpp \
+
+LIB=addon-callbacks-codec.a
+
+include ../../../../../../Makefile.include
+-include $(patsubst %.cpp,%.P,$(patsubst %.c,%.P,$(SRCS)))
+
