+++ xbmc/addons/binary/interfaces/api1/GUI/AddonGUIRenderingControl.cpp
@@ -1,0 +1,93 @@
+/*
+ *      Copyright (C) 2015-2016 Team KODI
+ *      http://kodi.tv
+ *
+ *  This Program is free software; you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation; either version 2, or (at your option)
+ *  any later version.
+ *
+ *  This Program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with KODI; see the file COPYING.  If not, see
+ *  <http://www.gnu.org/licenses/>.
+ *
+ */
+
+#include "AddonGUIRenderingControl.h"
+
+#include "guilib/GUIRenderingControl.h"
+
+namespace V1
+{
+namespace KodiAPI
+{
+namespace GUI
+{
+
+CGUIAddonRenderingControl::CGUIAddonRenderingControl(CGUIRenderingControl *pControl)
+  : CBCreate{nullptr},
+  CBRender{nullptr},
+  CBStop{nullptr},
+  CBDirty{nullptr},
+  m_clientHandle{nullptr},
+  m_pControl{pControl},
+  m_refCount{1}
+{ }
+
+bool CGUIAddonRenderingControl::Create(int x, int y, int w, int h, void *device)
+{
+  if (CBCreate)
+  {
+    if (CBCreate(m_clientHandle, x, y, w, h, device))
+    {
+      m_refCount++;
+      return true;
+    }
+  }
+  return false;
+}
+
+void CGUIAddonRenderingControl::Render()
+{
+  if (CBRender)
+  {
+    CBRender(m_clientHandle);
+  }
+}
+
+void CGUIAddonRenderingControl::Stop()
+{
+  if (CBStop)
+  {
+    CBStop(m_clientHandle);
+  }
+  m_refCount--;
+  if (m_refCount <= 0)
+    delete this;
+}
+
+void CGUIAddonRenderingControl::Delete()
+{
+  m_refCount--;
+  if (m_refCount <= 0)
+    delete this;
+}
+
+bool CGUIAddonRenderingControl::IsDirty()
+{
+  bool ret = true;
+  if (CBDirty)
+  {
+    ret = CBDirty(m_clientHandle);
+  }
+  return ret;
+}
+
+}; /* namespace GUI */
+}; /* namespace KodiAPI */
+}; /* namespace V1 */
