+++ xbmc/addons/binary/interfaces/api1/GUI/Makefile
@@ -1,0 +1,9 @@
+SRCS=AddonCallbacksGUI.cpp \
+     AddonGUIRenderingControl.cpp \
+     AddonGUIWindow.cpp \
+
+LIB=addon-callbacks-gui.a
+
+include ../../../../../../Makefile.include
+-include $(patsubst %.cpp,%.P,$(patsubst %.c,%.P,$(SRCS)))
+
