+++ xbmc/addons/binary/interfaces/api1/InputStream/CMakeLists.txt
@@ -1,0 +1,9 @@
+set(SOURCES AddonCallbacksInputStream.cpp)
+
+set(HEADERS AddonCallbacksInputStream.h)
+
+core_add_library(api1AddonCallbacks_InputStream)
+
+if(ENABLE_INTERNAL_FFMPEG)
+  add_dependencies(addons ffmpeg)
+endif()
