+++ xbmc/addons/binary/interfaces/api1/InputStream/Makefile
@@ -1,0 +1,7 @@
+SRCS=AddonCallbacksInputStream.cpp \
+
+LIB=addon-callbacks-inputstream.a
+
+include ../../../../../../Makefile.include
+-include $(patsubst %.cpp,%.P,$(patsubst %.c,%.P,$(SRCS)))
+
