+++ xbmc/addons/binary/interfaces/api1/PVR/Makefile
@@ -1,0 +1,6 @@
+SRCS=AddonCallbacksPVR.cpp \
+
+LIB=addon-callbacks-pvr.a
+
+include ../../../../../../Makefile.include
+-include $(patsubst %.cpp,%.P,$(patsubst %.c,%.P,$(SRCS)))
